requirements.txt
# Core requirements for the PureProt project
web3==6.0.0
python-dotenv==0.21.0
rdkit-pypi==2022.9.5
scikit-learn==1.2.2
pandas==1.5.3
chembl-webresource-client==0.10.8
tqdm==4.65.0


pandas>=1.3.0
scikit-learn>=1.0.0
rdkit>=2022.09.5
web3>=6.0.0
py-solc-x>=1.1.1
numpy>=1.22.0
matplotlib>=3.0.0

# For future integration with molecular modeling
# rdkit>=2022.9.5
# scikit-learn>=1.2.0
# torch>=2.0.0
